Implement basic entity factory interface
Entity factories map the descriptions of entities in your fixture
files to entities in your application. Implementing them can be
as simple as providing a _locate and _create method that knows how
to retrieve and create entities of your particular application type.

See the code and the Example factory for details.